+++ imap_tools/__init__.py
@@ -1,3 +1,6 @@
+# Lib author: Vladimir Kaukin <KaukinVK@ya.ru>
+# Project home page: https://github.com/ikvk/imap_tools
+
 from .query import AND, OR, NOT, Header, UidRange, A, O, N, H, U
 from .mailbox import BaseMailBox, MailBox, MailBoxUnencrypted
 from .message import MailMessage, MailAttachment
@@ -5,6 +8,4 @@ from .folder import MailBoxFolderManager
 from .consts import MailMessageFlags, MailBoxFolderStatusOptions
 from .errors import *
 
-__author__ = 'Vladimir Kaukin <KaukinVK@ya.ru>'
-
-__version__ = '0.45.0'
+__version__ = '0.46.0'